requirements.txt
notion-client
python-dotenv
pyyaml
requests
